requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
scikit-learn>=1.3.0
numpy>=1.24.0
langgraph>=1.1.0
langchain-core>=0.3.0
streamlit>=1.28.0
google-genai>=1.0.0
